重构和代码改进
- 重构TestAgent中的方法，将TimeWindow、Skills、Goals、Constrains和Workflow方法合并为一个SystemAgentOptions方法。
- 删除未使用的代码和注释，以提高代码清晰度和维护性。
- 在main.go中调整模型生成顺序，确保与业务逻辑的一致性。
- 启动Feishu客户端和异步任务管理器，增强服务的即时通信功能和后台任务处理能力。
- 删除无用的gen.go文件，以简化代码库并避免不必要的编译输出。

+++ model/daily_report.gen.go
@@ -10,12 +10,12 @@ const TableNameDailyReport = "daily_report"
 type DailyReport struct {
 	Sid        string `gorm:"column:sid;type:TEXT;not null" json:"sid"`
 	DataType   string `gorm:"column:dataType;type:TEXT;primaryKey" json:"dataType"`
-	UpdateTime int32  `gorm:"column:updateTime;type:INTEGER;not null" json:"updateTime"`
-	TimeInZero int32  `gorm:"column:timeInZero;type:INTEGER;not null" json:"timeInZero"`
+	UpdateTime int64  `gorm:"column:updateTime;type:INTEGER;not null" json:"updateTime"`
+	TimeInZero int64  `gorm:"column:timeInZero;type:INTEGER;not null" json:"timeInZero"`
 	ViewTag    string `gorm:"column:viewTag;type:TEXT;not null" json:"viewTag"`
 	Value      string `gorm:"column:value;type:TEXT;not null" json:"value"`
-	IsUpload   int32  `gorm:"column:isUpload;type:INTEGER;not null" json:"isUpload"`
-	IsDeleted  int32  `gorm:"column:isDeleted;type:INTEGER;not null" json:"isDeleted"`
+	IsUpload   int64  `gorm:"column:isUpload;type:INTEGER;not null" json:"isUpload"`
+	IsDeleted  int64  `gorm:"column:isDeleted;type:INTEGER;not null" json:"isDeleted"`
 }
 
 // TableName DailyReport's table name
